+++ src/devServer/devServer.ts
@@ -19,6 +19,7 @@ import {
 	getFileMimeType,
 	getFileContentsBuffer,
 	getFileStats,
+	getFileContentsHash,
 } from '../build/baseFilerFile.js';
 
 export interface DevServer {
@@ -36,7 +37,7 @@ export interface Options {
 }
 export type RequiredOptions = 'filer';
 export type InitialOptions = PartialExcept<Options, RequiredOptions>;
-const DEFAULT_HOST = 'localhost'; // or 0.0.0.0?
+const DEFAULT_HOST = 'localhost';
 const DEFAULT_PORT = 8999;
 export const initOptions = (opts: InitialOptions): Options => ({
 	host: DEFAULT_HOST,
@@ -101,7 +102,13 @@ const createRequestListener = (filer: Filer, log: Logger): RequestListener => {
 		// }
 		if (!file) {
 			log.info(`${yellow('404')} ${red(localPath)}`);
-			return send404(req, res, localPath);
+			return send404(req, res);
+		}
+		// console.log('req headers', gray(file.id), req.headers);
+		const etag = req.headers['if-none-match'];
+		if (etag && etag === toETag(file)) {
+			log.info(`${yellow('304')} ${gray(localPath)}`);
+			return send304(res);
 		}
 		log.info(`${yellow('200')} ${gray(localPath)}`);
 		return send200(req, res, file);
@@ -119,12 +126,17 @@ const toLocalPath = (url: string): string => {
 	return relativePath;
 };
 
-const send404 = (req: IncomingMessage, res: ServerResponse, path: string) => {
+const send404 = (req: IncomingMessage, res: ServerResponse) => {
 	const headers: OutgoingHttpHeaders = {
 		'Content-Type': 'text/plain; charset=utf-8',
 	};
 	res.writeHead(404, headers);
-	res.end(`404 not found: ${req.url} → ${path}`);
+	res.end(`404 not found: ${req.url}`);
+};
+
+const send304 = (res: ServerResponse) => {
+	res.writeHead(304);
+	res.end();
 };
 
 const send200 = async (_req: IncomingMessage, res: ServerResponse, file: BaseFilerFile) => {
@@ -132,10 +144,29 @@ const send200 = async (_req: IncomingMessage, res: ServerResponse, file: BaseFil
 	const mimeType = getFileMimeType(file);
 	const headers: OutgoingHttpHeaders = {
 		'Content-Type':
-			mimeType === null ? '' : file.encoding === 'utf8' ? `${mimeType}; charset=utf-8` : mimeType,
+			mimeType === null
+				? 'application/octet-stream'
+				: file.encoding === 'utf8'
+				? `${mimeType}; charset=utf-8`
+				: mimeType,
 		'Content-Length': stats.size,
-		'Last-Modified': stats.mtime.toUTCString(),
+		// TODO this works but not for source maps.
+		// not sure why. browser isn't returning the 'if-modified-since' header
+		ETag: toETag(file),
+
+		// TODO any of these helpful?
+		// 'Last-Modified': stats.mtime.toUTCString(),
+		// 'Cache-Control': 'no-cache,must-revalidate',
+		// 'Cache-Control': 'must-revalidate',
+
+		// TODO probably support various types of resource caching,
+		// especially if we output files with contents hashes.
+		// 'Cache-Control': 'immutable',
+		// 'Cache-Control': 'max-age=31536000',
 	};
+	// console.log('res headers', gray(file.id), headers);
 	res.writeHead(200, headers);
 	res.end(getFileContentsBuffer(file));
 };
+
+const toETag = (file: BaseFilerFile): string => `"${getFileContentsHash(file)}"`;